+++ core/src/test/kotlin/sift/core/api/testdata/set1/MethodWithParam.kt
@@ -5,6 +5,7 @@ package sift.core.api.testdata.set1
 class MethodWithParam {
 
     var helloField: Int = 4
+    var barField: String = "hi"
 
     fun fn(foo: Int, bar: String) {
         val a = 1